# test-microservices/payment-service/tempoutput/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Scan Results: payment-service",
    "version": "1.0",
    "description": "APIs discovered from payment-service (/code)",
    "x-generated-by": "Invicti AI API Scanner v1.0.0",
    "x-generated-at": "2026-03-04T11:28:28.193966",
    "x-service-name": "payment-service"
  },
  "paths": {
...
    "/webhooks/stripe": {
      "post": {
        "summary": "OpenAPI-Spec endpoint",
        "description": "Discovered from [openapi.json]:4200\n\nFramework: OpenAPI-Spec\nLanguage: Unknown\nRisk Level: LOW\nAuth Status: UNKNOWN",
        "operationId": "post_webhooks_stripe",
        "tags": [
          "OpenAPI-Spec",
          "Unknown"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "x-source-file": "/code/output/openapi.json",
        "x-source-line": 4200,
        "x-risk-level": "LOW",
        "x-auth-status": "UNKNOWN",
        "x-risk-reasons": [
          "Mutation: POST",
          "Shadow API"
        ],
        "requestBody": {
          "description": "Request body",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}